Update patch set 1
Patch Set 1:

I'd prefer to handle this a different way, one that doesn't require us to worry about whether the OS or the underlying filesystem supports fine-grained mtimes.  The feature where directory caching is really useful is the progress callback; it also won't break if it doesn't have absolutely the latest information.  Right now it's only called every 5 seconds but I want the ability to change this in the future. If we were to go to one second, the cache would always be bypassed if I read your code correctly.

I'm working on a patch (so this isn't a request for you to write one), but let me know your thoughts please.

Patch-set: 1