Update patch set 1
Patch Set 1:

> [...] Right now it's only called every 5 seconds but [...]

This does not reflect what I see, when running worker.py. To
show this, I added printing the current timestamp right before
?if (self.dirCacheOutdated(date)):?. This is my output:

  [...]
  Checking dircache: 1331375028.75
  Checking dircache: 1331375028.75
  Checking dircache: 1331375029.67
  Checking dircache: 1331375029.67
  Checking dircache: 1331375029.79
  Checking dircache: 1331375029.79
  Checking dircache: 1331375030.79
  Checking dircache: 1331375030.79
  Checking dircache: 1331375030.79
  [...]

Typically two successive checks followed by some pause.

> If we were to go to one second, the cache would always
> be bypassed if I read your code correctly.

This is not the way, I read my patch. But I am not sure,
whether I correctly understand your remark. Let me phrase
out in greater detail, how the patch works:

Assume, we now have timestamp 12.3, using ext3, and we detect
that the dir's mtime is 12 (the same integer timestamp) while
the cache's timestamp is 8.

The cache is outdated. We fetch a new cache.

The old approach was to set the cache timestamp to the dir's mtime: 12.
This is a problem. For example if at 12.5, a fileX is generated in the
dir. The dir's mtime is still 12 (ext3 cutting off .5 from 12.5). Hence if
we check again at 12.7, the dir's mtime (12) still matches the timestamp
stored for the cache (12). Hence the cache is marked as accurate, while it
is not accurate. We would not detect fileX. False cache hit.

What the patch does, is getting rid of those false cache hits by marking
only those caches as available for later use, that have been pulled for a
dir mtime, that has a different timestamp than a modification made right
now would yield.


So the patch works as follows:

* Assume, we again have timestamp 12.3, using ext3, and we detect
that the dir's mtime is 12 (the same integer timestamp) while
the cache's timestamp is 8.

The current dir's content is fetched (not from cache, but from live)
and returned.
As we saw above, we must not reuse the cache, as a later modification
(until 12.99...) would generate the same mtime. Hence, we would not be
able to detect dir modifications.
Hence, we set the cache's timestamp to 0, to mark that the current
cache shall not be reused.

Now, again assume that at 12.5, a fileX is generated in the
dir. Again, the dir's mtime is still 12 (cutting off .5 from 12.5).
But here is the difference: If we check again at 12.7, the dir's mtime
(12) does /not/ match the cache timestamp (0). Hence we recheck
the directory. This is necessary and we detect the new fileX.
No false cache hit.

So the original problem is fixed.
But does the cache still work?

Assume we check for files in the dir at timestamp 13.2. The dir's mtime
is still 12. So the patch's if condition
  13.2 >= 12 + 1:
is met and we store the dir's mtime 12 for the cache time.

If we check again at 13.4 the cache is used.
If we check again at 15.2 the cache is used.
If a file is added at 15.7, the dir's mtime becomes 15,
and subsequent checks detect that the cache is outdated.

So the patch eliminates false cache hits, but introduces false
cache misses.
However, those cache misses are essential. In the situation described
above, the filesystem /alone/ does /not/ allow to detect whether a cache
is still accurate or not. So we have to assume it is no longer accurate,
to avoid errors.

Patch-set: 1