Update patch set 1
Patch Set 1:

the progress callback is only called once every five seconds. the directory lookup is called in "batches" as you noticed from printing out timestamps; various routines that list files are called in clumps, which is why you see that behavior.

I'm only interested in caching for the progress callback, where false cache hits are not a concern as long as the cache is updated within a short period of time; I prefer it to having a large pile of cache misses in the mix.

Let me test my code and post it so you can see where I'm going with it and comment further.

Patch-set: 1