Update patch set 1
Patch Set 1: (1 inline comment)

Reading the patch I seem to realize that you never wanted the cache
to hold accurate information :D
The cache was just meant as hack for the status generation?

Ok, I can live with that.

Your patch certainly solves the original problem of false cache hits
when finding files to recombine.
So I would certainly love to see it committed. Yay!


However, it comes at a high price: It makes source code considerably more complex. 
This is a burden for new people reading the code.
This is a burden for us when rereading the code after 3 months.
And in some neglectable teeny weeny bit, we are affected by it already right now.

The Status class is documented to hold "everything that has to do with
reporting the status of a piece of a dump"

Now we have two times "everything around status". One time with
cache. One time without cache.

Those objects are /not/ kept in sync. So the natural question is, what
property is accurate in which of the Status instances.

For example failCount is only valid in Dump.status, not on
Dump.statusWithDirCache.
This might become a problem if for example anything starting from
Status.updateStatusFiles tries to flag errors or depend on the value
of failCount.
For example right now Status.updateStatusFiles calls
Status._saveStatusSummaryAndDetail, which calls
Status._reportDatabaseStatusSummary, and
Status._reportDatabaseStatusDetailed, which both call
Status._reportStatusSummaryLine which uses Status.failCount.

Hence, during a dump's progressCallback, status.html might be
outdated or present false status, as it relies on
statusWithDirCache.failCount, which should not be used. Only
status.failCount is kept current.

Right now, this is only a minor issue, as the Dump object separately
signals this failure to upstream and finally, the HTML gets updated
correctly.

But this flow of control illustrates how the second status object
affects us already now--although this does not hurt us immediately
now, as your original code provides a safety net for the status.

However, now we know the code, we know where to look.
What about finding such twisted flow of control and detecting such
problems in 3 months?

So I asked myself, if the two status objects are worth making the code
less readable.

In the proposed patch, caching is generally turned off. Caching is
only used in Dump.progressCallback.

However, Change-Id I17dc1abac0f3b1aca6646055b70895d7999b3224
also takes advantage of the cache in Dump.progressCallback. After 1 second
to settle the dir mtime, cache kicks in and it has /only/ cache hits within
Dump.progressCallback.
This first second is time-wise the difference between the two approaches.
On my machine a cache miss is roughly (rounded up) 0.0003 seconds. This
times 27 (enwiki), is 0.0081 seconds per /parallel/ job. For serial jobs,
the difference is even smaller.

To me, 0.0081 seconds (for a job that runs hours or days) is not worth
the trouble the second status object will bring in the long run.

Although both patches solve the problem, I'd nevertheless strongly
favor to commit Change I17dc1ab...
It solves the problem as well.
It is just as fast.
It transparently handles caching in DumpDir. The complexity of caching
it is /completely hidden/ within DumpDir, as it should be.
The patch is simpler, and less invasive.
The patch comes without hard to spot side effects (as discussed above)

Patch-set: 1