Update patch set 1
Patch Set 1: Abandoned

You are right.
It is more defensive to keep the try catch block.

So I separated the two different parts of the patch, and put first one (opening guard) into
  Change-Id: I3ff5a3c6cd62c2aff434892f6e2d8f1f9e730b13

An updated version of the second part (that now keeps the try catch block) is available at
  Change-Id: Ic56d883f4a396f758db8acbd52e30927919c74f6

Patch-set: 1