Update patch set 1
Patch Set 1:

Although after the previous commit lastFailed again does what it is supposed to do, checking dependencies via lastFailed is troublesome, as (for complete runs) it relies on the jobs being added immediately after the job it depends on.

This is likely to break in the future, if one adds further jobs later on. 
This has already happened for example for "articlesmultistreamdump", which ? through lastFailed ? relies on the 7z dumps "metahistory7zdump"/"metahistory7zdumprecombine" for complete dumps, while it should rely on the articlesdump(recombine).

As the each of the items using lastFailed already carries the item they depend on, I directly added a check for their status (This checking is refactored in a follow up commit).

Patch-set: 1