Update patch set 1
Patch Set 1:

Dependency information about the jobs was stored in two places:
- checkJobDependencies of DumpItemList
- __init__ of DumpItemList (initialization of self.dumpItems)

We wanted to get rid of this duplication.

The dependency information in the __init__ part is essential, as the items need it to determine the input files.

Therefore, we removed checkJobDependencies, which was used only for "--job" runs anyways.

Now all checking and specifying of dependency information is done uniformly, regardless of whether one starts complete dumps or "--job" runs.

Patch-set: 1