Update patch set 1
Patch Set 1:

While moving the check for Maintenance mode around in commit
c483a067b9eb3214ee1309ef2602bfb3c4c831f2 , the if guard's condition
for whether or not an item has to be run has been negated, but only
part of the while loop's branch has been re-indented.

So for items that having toBeRun false, the 'item.status() == "done"'
condition has been evaluated, and as such items need not be in state
"done" but may in status "waiting" (E.g.: stopping a dump and
afterwards running a specific job)
  self.runHandleFailure()
was called for such items. Hence it would wrongly flag dump failure.

The proposed patch fixes this issue.

Patch-set: 1