bugfixes: handle deleted text; workaround dupl text ids
Revisions with deleted text will get text id of 0 and len of 0
The rest will get a text id identical to the rev id to avoid collisions:
in live mediawiki projects it is possible for two revs to
have the same text id (example: revision created by protecting
a page), but rev ids are guaranteed to be unique.

If the 'nodrop' option is supplied, which means that
INSERT IGNORE statements will be generated for these
tables, the original text ids from the xml will be used.

That last workaround needs to be documented better. Way better.

Change-Id: I8ca20d7fb7c65713c7b62b74c963d2f298953932