Update patch set 1
Patch Set 1:

Thanks for this.
I've not tried rzip, but recently I played with lrzip (https://wiki.archlinux.org/index.php/Lrzip ) and I saw it can consume really titanic amounts of memory (as much RAM as the data is, or a lot of disk writes with -U option): is rzip the same? If you remember, it would be nice to write down how much memory it consumed on that test instance and how it scaled to much bigger datasets.

Patch-set: 1
Reviewer: Nemo bis <105@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0