Update patch set 2
Patch Set 2:

Tried and failed to submit this as a comment on gerrit, so here it is over e-mail:


1. Ariel's comments--benchmarks.

I don't have it down to a standard script, but the numbers in README were from running on dumps from the archives--I just looked at whatever was the latest full-history dump at the time and pulled one of the smaller files, then ran the scripts on it on an EC2 instance.

For blks-standalone.py, that was:

  In a test on a c1.medium EC2 instance, it compressed 4.1G to 63M in 2m45s
  wall time. Using bzip instead of gzip on the text, it gets the file to 56M
  in 3m40s.

blks-rzip.py did a little better:

  in the [same] test scenario above, it creates a 44M
  file in three minutes.  

I didn't capture RAM use, but I don't remember problems with it; the script compresses each 10MB block independently, so there's no reason for it to allocate GBs.

Some other numbers from all the playing around are at http://www.mediawiki.org/wiki/Dbzip2#rzip_and_xdelta3.


2. Nemo bis's comments--rzip RAM use.

rzip is like lrzip -U: it doesn't keep the file in RAM, but instead goes back to the filesystem when it finds a match with old text. So ballooning RAM wasn't a problem in my tests, but general inefficiency still is. That's one reason blks-rzip.py split the file into 10MB blocks first (the other reason was to allow quick-ish random reads).


3. General thoughts + other code.

I think delta coding (which I think Svick plans to integrate into his new dump format) is the Right Way to attack this problem. Unlike rzip, a delta coder 'knows' to look for matches against just the last rev or two, not the whole file. That can save lots of work and RAM.

A long while back I pointed Svick towards some pre-existing delta-coding libraries he might be able to use: http://www.mail-archive.com/xmldatadumps-l@lists.wikimedia.org/msg00495.html. Still think that's a good idea.

I also wrote a delta coding program in Go, not for compressing fulldumps, but for diffing one monthly pages-articles dump against another, or a monthly dump against the last revision from an incr. That's at https://github.com/twotwotwo/dltp and a post to wikitech-l about it is at http://lists.wikimedia.org/pipermail/wikitech-l/2013-August/071174.html. It's been dormant for a bit, and the programming-language mismatch means it's probably not that useful to the C++ dumps project. But it's out there.

Patch-set: 2
Reviewer: Twotwotwo <819@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0