Update patch set 1
Patch Set 1:

What is the status of this very old changeset? Should it be abandoned? Or at least given a -1? Or is it work in progress?

operations/dumps now appears at the #5 position in the ranking of slowest projects reviewing contribution, but it looks like this is caused by some patches that nobody is working on or reviewing nowadays.  http://korma.wmflabs.org/browser/gerrit_review_queue.html

Patch-set: 1
Reviewer: Qgil <465@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0