Update patch set 1
Patch Set 1:

(2 comments)

It looks like you rewrote much of dump_restbase in python ;) I thought that we'd just create a script to set up the directories & do some tweaks in dump_restbase, but this is fine too, especially if it's easier to maintain for you.

The naming looks fine to me, only comment is about using full timestamps.

Things that are missing:

- Parallelism. With dump_restbase a parallelism of 2-3 makes a big difference.
- Compression and decompression, see dump_restbase.
- Copying of previous dumps into the workdir, which is needed for incremental dumps. A full dump is simply a dump without an existing DB.

Patch-set: 1