dumps: redo handling of jobs with unrun prereqs
for a wiki not currently locked by another worker, if a prereq for a
job is waiting or in progress, don't call this run a failure,
just mark as waiting.  prereqs that failed still result in this
run being marked as failed

if we have requested specific job(s) and skipdone:
for a wiki locked by another worker, if a prereq job is waiting
or in progress, we won't return "no wikis available
to run" as that won't strictly be true. We're just too early.

fix couple typos in comments

for createdir job, generate dumprun info file as for most jobs

Change-Id: I8cf8d27dda0d14262a39c56ac73f7bc00005719a