dumps: clean up history dumps code, dump only missing pages for reruns
This large refactor simplifies the code path for generating history
dumps (and there is still more to do), so that dumps with or without
checkpoint files or subjobs all use the same methods.

If this is a rerun from a previous run, the old output files will
be checked and only commands to dump page ranges missing from those
will be generated.  This includes generating partial stub files
for only those page ranges.

We don't check to see if the old output files are truncated; they
are presumed good as having survived a previous run.  If they are
bad they will be moved out of the way on this run and the run
itself will abort.  A subsequent run would then cover missing
page ranges for those files in addition to whatever else might
be missing. (Might change this in the future.)

This refactor relies on the cleanup of files from the old run
not happening.  An option should be added to the script to
switch old file cleanup on or off, and the cleanup itself needs
to be fixed up, as it does not work properly for checkpoint files.

Change-Id: Ib5389f012ddf7a4dcf20e6bbdacf4a9c5b59ecdd