dumps mirroring tool, don't assume dest is local filesystem
This was a holdover from the daze when it was used to copy files
to glusterfs locally nfs mounted.  Those days are gone, coffin
nailed shut, weighted down with concrete and tossed into the ocean.

This will do batched rsyncs of dump files from a list, from source
to dest, either may be remote, and will do cleanup on the dest of
directories per wiki project which were not copied (assumed to be obsolete).

Change-Id: Ia4a1f01c0aaef273a1191d4a64d451f0f2a87e38