script to run a maintenance command on all wikis with varying output dirs
Results can go into a directory that varies by date and wikiname as
well as output files that may include the date, wikiname and script
name.  Retries may be specified.

This script skips runs on closed or private wikis.

Config files used are the same as the wiki dump config files.

Change-Id: I0cafec1678cf6ac0f6db6b499e764fbbea6cae9a