add option to onallwikis to run from a base wiki, with wikis as args
The idea here is to be able to run a query or a script that takes as
in its arguments a wiki name which needs to loop through all wikis
in some list, while the script/query itself runs from one base wiki.

This might be true for e.e. commons (global usage data) or wikidata.

Also fix typo in usage message and move a section for clarity.

Change-Id: Iad40e440af5d919b67418bdc9a00eb0a2e958efb