lengthen max sleep time between jobs and randomize
too many jobs running at once means too many contenders for the lock
for the last wiki with a job yet to run, when the rest of the runners
have moved on to the next stage and find missing dependencies.

randomize the sleep time as well so that every job runner will get a
chance at the wiki, including the runner that is processing the current
job

This resolves issues such as the following: One wiki does not yet have
the articles dump run, all others have it completed and the recombine
phase has completed for all wikis except for the one waiting for
articles.  The runner doing the articles dump can never get the lock
for the one wiki remaining because one of the other runners always
gets there first.  Thus no remaining jobs are ever processed.

Change-Id: Id02287e3d898c72ff07bb604a519eb8b9fe45376