add ability to do xml stubs dump pieces based on revs per page range
Generate crude estimates that won't burden the servers, better
to have more lighter queries than fewer and have them be horrid
temp filesort things.

This is needed because we are now explicitly ordering revisions
within each page for stubs and for content.  The query for that
ordering will be prohibitively expensive if the number of revisions
for the page range is too high.

Bug: T137887
Change-Id: Id33df29e1830ca53a6f782b4cffe04249d3ad9e1