allow list of jobs to run to be passed as argument for dump scripts
previously this option permitted only one job to be specified.

This will allow simplifcation of dumps stages runs and will
avoid the problem where failure on one job in a stage for one
wiki means that all runners block on the next dependent job,
if these two jobs are listed in the same stage.

With this change, the entire job sequence will be run by each
runner until there are either too many failures or the failed
job completes.

Change-Id: I43c236033791562399e0b136aa4508e80599eeab