Make scheduler hupable.
HUP will have scheduler shoot all children and restart itself.  All threads
set to daemon so they will terminate on parent exit; setpgrp used in the
dump scheduler so all children in the tree will receive kills; convert
any users of Thread module to users of Threading module for consistency

Not sure if, on reload, we want scheduler to toss the run cache or not;
for now leaving the cache alone to be reused on restart if rerun was
in the starting args.

Bug: T142488

Change-Id: I15adf01232c1774d53ebcd38d284706387e1d508