scheduler: move generation of unique id string out to the module
We need this to get called once only, so it's consistent across
the entire script run.  It therefore makes sense for the same code block
that instantiates the Scheduler to call this first, which means it
should live outside of any class.

also fix for real the file closing on HUP so we actually check that
we are not closing the descriptors for stdin/out/err

Change-Id: Ia341cc0c61a7eacd195dd92cffcf5516150f0496