Update patch set 1
Patch Set 1:

Well after a ridiculously long time I am ready to make this happen.  We need to think about where the defaults file should live (hardcoded path? relative to what?)  This is part of the dumps/ subdirectory now.  It goes against my grain to have a text file in there with the python modules; is that silly?

Patch-set: 1
Label: Code-Review=0