move last references to incr/Incr out of generateincrementals module
there is now a miscdumpsfactory module; any new misc dump types
should be added to this in the manner of the incremental dumps.

to generate a dump one must now specify a dumptype; currently only
incrementals are supported but soon html dumps will be added to
the list.

updated README.incr.txt to reflect new files

Bug: T133547
Change-Id: I057c60ee0982820828c0088d13a3558659c6e2e1