fix prefetch setup for retries of content file dump steps
with checkpoint files enabled, the temp file to be written for a
retry, covering only partial page ranges, does not have leading
zeros for the page ids in the file name, the prefetch files will.
this means the comparison between values was broken, resulting
in no prefetch for this specific case.

Change-Id: Iaf02f9ae772d91dc46c39fa66b8c6c5ab54ab9a1