config setting to permit a list of wikis to be dumped in a specific order
We don't want these for all wikis but it can be useful for small lists.

Example the big wikis that dump via 4 processes at a time.  Some of these
take days longer than others; if we start those first, multiple other wikis
will run to completion on other cores, while the first ones chug along.
If the order is by longet to dump, these big wikis often wind up
starting near the end ofthe list, and run along by themselves after
everything else is completed.

We can't really just expand the configs of these big slow wikis
so they use a pile more processors at once, that's not ok for the db
servers. (enwiki, wikidatawiki, yes, but not the rest.) So, move them
to the front of the queue.

Change-Id: I494ed57363b1ddfe0e10be0aed25facb7ca8a364