Update patch set 2
Patch Set 2:

This function is only called if the config file doesn't have entries in the [database] section for 'user' and 'password'. It's specifically a fallback for that case.  Perhaps something else is wry in the vagrant setup?

Patch-set: 2
Reviewer: ArielGlenn <6@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0