don't rely on grep trick to see if there's binary crap in xml files
What we see when there's crap is a bunch of nulls right after the
header, so look for those instead (string of ^@) and whine if there
are any in the first so many lines.

All these checks need to be done on the final file location, since
the file is moved from temp name to final name before the checks
are made.

Change-Id: Ia6624dd6b20bab7fc8bd5afaf6c26348a052890c