generate checksums on a per job basis, updating the hash as needed
Right now, on every run of any job we regenerate all hashes of all
output files, appending each one to a temp file as we go.
Instead, write each hash to a per-job file, as well as to a
cumulative file, but if the per-job file already exists and is
more recent thaan the output file, just re-use the contents.

This will save a lot of extra processing time on some of the big
wikis.

Change-Id: I66a4cb2069aae0eab13a328624d84b75dca08a74