split up dumps temp dir into subdirs
When you have a lot of wikis, and for some things I want to start
putting in the temp dirs for dumps, we're going to wind up with a lot
of files. Even just for human eyes, it will be nicer to split
then up into subdirs, tempdir/a/aawiki  and so on.

Bug: T182572
Change-Id: I3d5306dbc30b4a958f864d338eccb3de94b6ea57