remove stuff now in the obsolete branch
There's some things in tools that are in the obsolete branch
because I forgot and left them there; once those are cleaned up,
the only copy of things we are interested in and might update
once in a while, but not useful enough to be in master, will be here.

There were some files that are in other repos (eiher puppet
or mwbzutils) so I tossed the copies of those from here as well.

Also had to update tox config to use python 2.

Change-Id: Ib5a159ca84e8e750d313dbb1a3553ee87971ecb2