pylint, also fix up failure error message for misc dumps
Although the third item in the arg for concat is of type string,
it isn't handled in the same was as the two string literals for
purposes of implicit concatenation. Be explicit instead.

While we're here, move the logging setup stuff from main to a
separate method, and get rid of a 'check this' comment that
is unneeded

Change-Id: I60a13a54a0bb6c0ed2f1d2774d2e236c142ac094