use MediaWiki maintenance script to get db user and password
If they aren't specified in the config file, do NOT parse some
horrible php file for the values, make MediaWiki do the work, since
it already can.

This entails moving the values to the Wiki object, so that
the db user and password can vary by wiki.

Parsing per project was also added to the Wiki object so that
it need not be done separately by the caller.

The miscdumps library has also been updated similarly.

Bug: T218923
Change-Id: Icd828d2d5956fdbe47ccb28d9571ef5eef34bd6c