generate index.html file for incr dumps once per pass over all wikis
Doing it after each wiki is too slow with nfs and a large directory
of wikis to process.

We add a sleep after each wiki which wil be reduced over the next
few days in subsequent patches, so that we can gradually speed up
the incrementals run such that runs finish in calendar order.

Bug: T221515
Change-Id: Iab0dee23b8fd8d0ae9ef39fb22c96e4bd43e75a2