split up page content jobs with max bytes per page range
The issue here is that there are now some pages on one of our projects
that, uncompressed and with all the revisions, are over 1GB, and so a
few hungred of those get very large and slow when dumping. This new
config option allows us to split jobs into page ranges that, in addition
to not having too many revisions, also don't have too many uncompressed
bytes. Better a bunch of tiny jobs (that can be run in parallel) than
a few large ones.

There was also a pylint whine about a too-short var name so that's
fixed too.

Bug: T221504
Change-Id: I5d956251fe98cdbd91c908c5559842e852aba410