one-off for generating some page meta history files
This is intended to be used for catch-up purposes, when the input stub
files for the various page ranges have already been generated (and
most likely the metahistorybz2 job is already running on some other
host).

You can specify the part numbers and page ranges so that any missing
output files can be regenerated in batches from other hosts.

Change-Id: Ib1c809e4b9d429ecae28f6b508335c8f7ca902ee