ability to configure a wiki to produce empty abstract files
Some wikis such as Wikidata have no content in the main namespace that is
abstractable. Instead of requesting content for the latest revision
for all of these pages only to discard it, just produce an output
file that has the start and end feed tags, and that is all.

Also fix up gz recombine jobs so that input files with start and end tags
and no content do not break but recombine properly.

Bug: T236006
Change-Id: I7bfff93ed9c059b3085c9f028dc02e7ae06f76fb