for 7z production in batches, skip files that exist at beginning of each batch
The idea here is that we can (and do) check for existence of the output files
at the beginning of the entire 7z recompression job. But in our glorious
future of multiple hosts doing recompression for the same wiki, or even
in our inglorious present of running catchup scripts on other hosts manually,
output files might land after the job has started. We want to make sure
we don't clobber them.

Also added a unit test to check that 7z commands are properly
generated for page content files, with or without pre-existing
7z files created between batches.

Bug: T250260
Change-Id: I5268ca7c817b8b683c5faa5351496b0eab9a5050