properly handle failure of writing of temp stubs for page content jobs
Previously we could not get the exit codes from failed pipelines in
a list of command_series, and so we could not properly check them
to make sure none were sigpipe and ignore. Now we'll skip flagging
those as errors properly.

This also fixes up checking for specific signal values, which changed
from Python 3.4 and below to Python 3.5.

Unittests are added for the new sigpipe checking.

Bug: T242209
Change-Id: I1b4b7eaffaded5465e7d5332694036ffc8876cf2