rename 'parts' attribute of Dump subclasses to something more accurate
'parts' in all the file listing methods is a list of part numbers.
The same argument in the Dump subclasses means something unrelated
and it was easy to confuse the two.

Also change the default value from false to None, so we don't have
a parameter that takes boolean and list values.

Bug: T246465
Change-Id: I98c5c5e0152056de4b0aa9fd500380f536e447c8