New class for output file listing methods to move them out of jobs code
Move all output file list methods in Dump subclasses into their own classes.
They are now indirect subclasses of JobFileLister, so all of the listing
methods are now separated from the various dump job classes to the
extent possible.

Also refactored a few methods of Dump subclasses to convert them to
staticmethods so that they could be called directly from the file lister
class

And finally, converted 'args' to a named tuple, got rid of the calls to
set defaults for the values of the tuple, and stopped modifying 'args'
from within a method to which it's passed.

Bug: T246465
Change-Id: If824482d98e7c66c9c1f8b1f44ff378ed58eb689