remove capability to dump private tables
If someone wants to dump private tables, this can be done via a separate
config file and yaml file of tables, writing only to a private location
including checksum files and so on.

For now we continue to check the 'type' param in the yaml file if
present, in which case it must have the value 'public' to be dumped.
If the param is not present, the table is deemed to be dumpable.

Bug: T249508
Change-Id: I890a34af8bf4237a7c9ccbccb19dfd7cb6c4d247