read special files from directory with the correct date
In the case where parts of a previous date's dump are being rerun,
this could cause some reads (but not writes) of special files
such as the pageranginfo file to fail, because they tried to
find the file in the directory for the most recent run, whether
it existed or not.

While this hasn't impacted any recent runs, it did make some testing
of (removal of) private table dumps fail.

Bug: T249508
Change-Id: Idd672081eb09a00ba0deb5133c40eec3fc8ff209