add the ability to skip a job via configuration
This is separate from the skipjobs option to worker.py
Skipping a job via configuration means that the job is not considered
to be part of a dump, and a dump run that completes having skipped the
specified jobs will be considered fully complete, not a partial run.
Jobs skipped in this fashion will not be displayed as skipped in the
generated index.html file, nor will they be reported on in any way.

This facility can be used to permanently skip sample jobs that are
added for purposes of example.

Some extra checks were added in the construction of the dumps item list,
because any job might be missing (skipped).

Change-Id: If0472e50914c49ca8c9a95d4ff9ae40c40f30778