Prevent mysql passwords from being logged to stdout
I've spotted mysql passwords in the stdout of the
xmldumps-backup/worker.py output. As we migrate the dumps to airlfow,
these will then be logged on the stdout of a container, which is then
sent to logstash, displayed in the airflow UI as well as stored in S3.

We'd like to redact these passwords from stdout.

Change-Id: I4535f3eef7633a41565e1a8332a0a7eda90ea23d
Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T388378