Add a copy of the dump scripts that are in puppet
We are going to need these scripts in order to carry out many of the
wikibase, cirrussearch, and other miscellaneous dumps.

They have previously been deployed to the snapshot hosts by puppet, but
now we need them to be in the mediawiki-cli container that will be used
for executing dumps.

These scripts have been copied verbatim from puppet, with the exception
of removing the header that says they are managed by puppet.

Bug: T394389
Change-Id: Ib51bdf870983408cddb633c208b7c9b23739386c