Adapt dump scripts for running in containers
The wrapper scripts that we recently copied into this repository from
puppet have some elements related to log handling which make them less
suitable for running in a container environment.

This commit updates the scripts to remove piping output to local log
files and the use of `logger`, which requires `/dev/log` to be usable.

This change will not affect the existing snapshot servers and the
production dumps, since they are still managed in puppet. These changes
will only be made in the mediawiki-cli image that pulls this dumps
repository.

Bug: T394389
Change-Id: Id2fd54b6e1692c51578d444a40bdb74036434132