Use 'set -o pipefail' instead of 'set -e' in wikibase scripts
We have previously attempted to address reports of corrupted wikibase
dumps by adding 'set -e' to the top of the wikibase dump scripts.

Unfrotunately, this made the scripts brittle in unexpected ways and this
has resulted in some failed wikibase dumps, requiring back-fill.

We believe that the cause of the corrupted files was that several of the
chained commands are not within the scope of the existing 'set -o
pipefail' statements.

These are particularly within the section of the script that carries out
the compression.

In order to address this, we set the 'set -o pipefail' option at the
top of the script and remove the 'set -e', for now.

Bug: T400383
Bug: T399077
Bug: T399119
Change-Id: I118f9989d1f5491cc904913253359701c4110265