Use relative path for "latest" symlinks
The "latest" symlinks for the wikidata entity dumps broke due to
Ie7047c12e29480702a818c8d8aa6207821359865 creating symlinks with
absolute paths. The link targets point to the wrong location after the
dump files are synced to the public dumps server.

This change makes use of the -r option to ensure the symlinks are
created with relative paths which stay valid on the dumps server.

Bug: T412726
Change-Id: I647507ddfdc3afef529a63aada424dd6a8a20725